day 12 part 1 (im giving up)

+++ app/src/main/java/com/example/advent/day11/Day11.kt
@@ -54,8 +54,8 @@ private fun calculateStones(): Long {
     return result
 }
 
-private fun replaceStonesRecursion(blinks: Long, number: Long): Long { // recursion and memoization
-    if (blinks == 0L) { // base case
+private fun replaceStonesRecursion(blinks: Int, number: Long): Long { // recursion and memoization
+    if (blinks == 0) { // base case
         return 1
     }
     val key = "$number at $blinks"
